+++ src/Types/Graphics/Color.cpp
@@ -1,7 +1,0 @@
-#include "Color.h"
-
-namespace Types {
-namespace Graphics {
-// Nothing.
-}
-}