+++ src/Types/Graphics/Point.cpp
@@ -1,7 +1,0 @@
-#include "Point.h"
-
-namespace Types {
-namespace Graphics {
-// Nothing.
-}
-}