+++ src/love/Types/Graphics/Color.cpp
@@ -1,0 +1,11 @@
+#include "Color.h"
+
+namespace love {
+namespace Types {
+namespace Graphics {
+
+// Nothing.
+
+}  // namespace Graphics
+}  // namespace Types
+}  // namespace love