+++ src/love/Types/Graphics/Image.cpp
@@ -5,9 +5,10 @@
 
 #include <iostream>
 #include <string>
-#include "../../ChaiLove.h"
+#include "../../../ChaiLove.h"
 
 
+namespace love {
 namespace Types {
 namespace Graphics {
 
@@ -85,3 +86,4 @@ int Image::getHeight() {
 
 }  // namespace Graphics
 }  // namespace Types
+}  // namespace love