+++ src/love/Types/Input/Joystick.cpp
@@ -1,8 +1,9 @@
 #include "Joystick.h"
 #include "SDL.h"
-#include "../../ChaiLove.h"
+#include "../../../ChaiLove.h"
 #include <string>
 
+namespace love {
 namespace Types {
 namespace Input {
 
@@ -52,3 +53,4 @@ void Joystick::close() {
 
 }  // namespace Input
 }  // namespace Types
+}  // namespace love