+++ src/love/font.h
@@ -1,7 +1,7 @@
 #ifndef SRC_LOVE_FONT_H_
 #define SRC_LOVE_FONT_H_
 
-#include "../Types/Graphics/Font.h"
+#include "Types/Graphics/Font.h"
 
 namespace love {
 /**