+++ src/love/sound.cpp
@@ -3,7 +3,10 @@
 #include "sound.h"
 #include "SDL.h"
 #include "../ChaiLove.h"
-#include "../Types/Audio/SoundData.h"
+#include "Types/Audio/SoundData.h"
+
+using ::ChaiLove;
+using love::Types::Audio::SoundData;
 
 namespace love {
 
@@ -17,13 +20,13 @@ bool sound::update() {
 
 void sound::unload() {
 	// Unload all sounds.
-	for (std::vector<::Types::Audio::SoundData*>::size_type i = 0; i != sounds.size(); i++) {
+	for (std::vector<SoundData*>::size_type i = 0; i != sounds.size(); i++) {
 		sounds[i]->unload();
 	}
 }
 
-::Types::Audio::SoundData* sound::newSoundData(const std::string& filename) {
-	return ::ChaiLove::getInstance()->audio.newSource(filename);
+SoundData* sound::newSoundData(const std::string& filename) {
+	return ChaiLove::getInstance()->audio.newSource(filename);
 }
 
 }  // namespace love